requirements.txt
numpy
pandas
biopython
matplotlib
requests
sklearn
typing
biopython
tabulate
joblib